// apps/web/src/api/model/getTags200TagsItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Rootly API
 * OpenAPI spec version: 0.1.0
 */
import type { GetTags200TagsItemColor } from './getTags200TagsItemColor';

export type GetTags200TagsItem = {
  id: string;
  name: string;
  slug: string;
  color: GetTags200TagsItemColor;
  workspaceId: string;
  createdAt: string;
};